Добавил Отдельные компоненты для тестов
