+++ examples/screenshot-test/brackets/EmptyFIlterWrapper.tsx
@@ -1,85 +1,13 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-import { FilterConditionEditorButton } from "../../../src/FilterConditionEditorButton";
-import { MenuMode } from "../../../src/FilterConditionSelector";
-import {
-	FilterWrapper,
-	ScrollState,
-	SelectionStateType
-} from "../../../src/FilterWrapper";
-import { FiltrationGroupComponent } from "../../../src/FiltrationGroupComponent";
-import { LegacyFilterGroupButton } from "../../../src/LegacyFilterGroupButton";
-
-const ChildRenderer = () => <div />;
+import { FilterWrapperTest, FiltrationGroupComponentTest } from "../utils";
 
 const EmptyFilterWrapper = () => (
-	<FilterWrapper
-		scrollState={ScrollState.Full}
-		filterActions={[]}
-		statisticsDescription="Всего клиентов"
-		statisticsValue={1021318}
-		doesContainFilter={false}
-		selectionState={SelectionStateType.All}
-		selectedText="Выбрано"
-		selectedCancelText="Отменить"
-		selectedCountDescription="100 (все)"
-		applyButtonCaption="Применить фильтр"
-		clearButtonCaption="Сбросить фильтр"
-		buttonUpCaption="Вверх"
-		filterActionsCaption=""
-		isDataOutdated={true}
-		onApply={() => {}}
-		onClear={() => {}}
-		onCancelSelection={() => {}}
-	>
-		<FiltrationGroupComponent
-			state="view"
-			groupType="or"
-			andLabel="И"
-			orLabel="ИЛИ"
-			shouldShowLabel={false}
-			shouldShowButtons={true}
-			shouldShowDuplicateButton={false}
-			addSimpleConditionButton={
-				<FilterConditionEditorButton
-					label="Добавить фильтр"
-					toggleOpen={() => {}}
-					isOpened={false}
-					iconType="filter"
-					searchTerm=""
-					onSearchTermChange={() => {}}
-					childRenderer={ChildRenderer}
-					rootIds={[]}
-					filterLabel=""
-					recentLabel=""
-					savedLabel=""
-					examplesLabel=""
-					onModeChanged={() => {}}
-					menuMode={MenuMode.Filter}
-					helpCaption=""
-					helpComponent={() => {}}
-					editorComponent={() => {}}
-					starred={false}
-					onConditionStateToggle={() => {}}
-					onPreviousSelected={() => false}
-					onExpandCurrent={() => {}}
-					onNextSelected={() => {}}
-					searchPlaceholder=""
-				/>
-			}
-			addGroupConditionButton={
-				<LegacyFilterGroupButton icon="cross-arrows">
-					И
-				</LegacyFilterGroupButton>
-			}
-			onGroupTypeToggle={() => {}}
-			onConditionStateToggle={() => {}}
-			onConditionRemove={() => {}}
-			onConditionCopy={() => {}}
-		/>
+	<FilterWrapperTest>
+		<FiltrationGroupComponentTest shouldShowButtons={true} />
 		<i>Добавьте фильтр, чтобы создать выборку клиентов</i>
-	</FilterWrapper>
+	</FilterWrapperTest>
 );
 
 const entryElement = document.createElement("div");
