+++ examples/screenshot-test/utils/LegacyFilterGroupButtonTest.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { LegacyFilterGroupButton } from "../../../src/LegacyFilterGroupButton";
+
+export const LegacyFilterGroupButtonTest = () => {
+	return (
+		<LegacyFilterGroupButton icon="cross-arrows">И</LegacyFilterGroupButton>
+	);
+};
